Move entity acceptance tests to parent package to fix CI 'Invalid resource type' error
The CI acceptance test runner was reporting 'Invalid resource type' when running
tests from the entity sub-package. Moving acceptance tests to the existing
security_entity_store_test package (as external tests in acc_test.go) ensures
the test binary is built correctly with all provider registrations.
